UI changes

+++ app/src/main/java/com/jagadeesh/thoughtworks/modal/WeatherInfo.kt
@@ -42,6 +42,6 @@ data class Sys(
         val id: Int,
         val message: Double,
         val country: String,
-        val sunrise: Long,
-        val sunset: Long
+        val sunrise: Int,
+        val sunset: Int
 )
